pyproject.toml
[project]
name = "modal-apis"
version = "0.1.0"
description = "Modal-hosted Coqui TTS and OpenAI Whisper APIs"
requires-python = ">=3.12"
authors = [
    {name = "Abhiroop Prasad"}
]

dependencies = [
    "modal>=0.63.0",
    "fastapi[standard]>=0.108.0",
    "pydantic>=2.5.0",
    "coqui-tts>=0.27.3",
    "structlog>=25.5.0",
    "numpy<2.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["coqui_service", "whisper_service", "shared"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by formatter)
